adding location
