+++ client/src/pages/Home/index.js
@@ -1,1 +1,2 @@
-export { default } from "./Home.js";
+export { default }
+from "./Home.jsx";
