+++ package.json
@@ -34,6 +34,7 @@
     "start-server-and-test": "^1.9.0"
   },
   "dependencies": {
+    "@ishawnwang/withunmounted": "^1.0.1",
     "axios": "^0.18.0",
     "bcrypt": "^3.0.2",
     "bcryptjs": "^2.4.3",
